Fix misnamed packages
When the CPAN dist file ends with .zip, or anything other than .tar.*,
the package name generated would include the suffix.

For example:
- MP3::Tag -> perl-mp3-tag-1.16.zip-noarch-1h3x.txz
- Config::Tiny -> perl-config-tiny-2.29.tgz-noarch-1h3x.txz

This strips .zip, and .tgz from the package names, as well as the most
common .tar.*.

Further research would be needed to determine if there are other
formats in the CPAN collection.